+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Change history for ui-tags
 
 ## 1.0.0 (IN PROGRESS)
+
+* Add ability to enable tags app. Fixes FOLIO-1302.